requirements.txt
asgiref==3.3.1
Django==3.1.6
django-debug-toolbar==3.2
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
# For Linux/mac
psycopg2-binary==2.8.6
# for Windows just psycopg2
#
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
